Implement iterative solver for projection step
The new implementation iteratively solves for the projection on each element with the CG algorithm.
However, it turns out that CG has problem to converge a system with many small different eigenvalues due to roundoff.
Therefore, the implementation uses the inverse mass matrix as a preconditioner such that all
eigenvalues of the projection matrix that are in the null space of the divergence operator (i.e.,
divergence-free velocities) are equal to 1. This gives good convergence despite condition number of
1e6 or more.